* changed pk + fk in db to int(10)

+++ config/db/tables.sql
@@ -3,11 +3,11 @@
 --
 
 CREATE TABLE `ext_comment_comment` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `date_update` int(10) unsigned NOT NULL DEFAULT '0',
   `date_create` int(10) unsigned NOT NULL DEFAULT '0',
   `deleted` tinyint(1) unsigned NOT NULL DEFAULT '0',
-  `id_person_create` smallint(5) unsigned NOT NULL DEFAULT '0',
+  `id_person_create` int(10) unsigned NOT NULL DEFAULT '0',
   `id_task` mediumint(9) unsigned NOT NULL DEFAULT '0',
   `comment` text NOT NULL,
   `is_public` tinyint(1) unsigned NOT NULL DEFAULT '0',
@@ -25,8 +25,8 @@ CREATE TABLE `ext_comment_feedback` (
   `id` int(10) unsigned NOT NULL AUTO_INCREMENT,
   `date_create` int(10) unsigned NOT NULL,
   `date_update` int(10) unsigned NOT NULL,
-  `id_person_create` smallint(5) unsigned NOT NULL,
-  `id_person_feedback` smallint(5) unsigned NOT NULL,
+  `id_person_create` int(10) unsigned NOT NULL,
+  `id_person_feedback` int(10) unsigned NOT NULL,
   `id_comment` int(10) unsigned NOT NULL,
   `is_seen` tinyint(1) unsigned NOT NULL,
   PRIMARY KEY (`id`)
@@ -39,11 +39,11 @@ CREATE TABLE `ext_comment_feedback` (
 --
 
 CREATE TABLE `ext_comment_mailed` (
-  `id` int(11) NOT NULL AUTO_INCREMENT,
+  `id` int(10) NOT NULL AUTO_INCREMENT,
   `date_create` int(10) unsigned NOT NULL,
   `date_update` int(10) unsigned NOT NULL,
-  `id_person_create` smallint(5) unsigned NOT NULL,
+  `id_person_create` int(10) unsigned NOT NULL,
   `id_comment` int(10) unsigned NOT NULL,
-  `id_person_mailed` smallint(5) unsigned NOT NULL,
+  `id_person_mailed` int(10) unsigned NOT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=MyISAM  DEFAULT CHARSET=utf8;
